+++ Editor/Drawers/Wwise/WwiseAutoGetterDrawer/WwiseAutoGetterAttributeDrawerUIToolkit.cs
@@ -1,4 +1,6 @@
 #if WWISE_2024_OR_LATER || WWISE_2023_OR_LATER || WWISE_2022_OR_LATER || WWISE_2021_OR_LATER || WWISE_2020_OR_LATER || WWISE_2019_OR_LATER || WWISE_2018_OR_LATER || WWISE_2017_OR_LATER || WWISE_2016_OR_LATER || SAINTSFIELD_WWISE && !SAINTSFIELD_WWISE_DISABLE
+
+#if UNITY_2021_3_OR_NEWER
 using System;
 using System.Collections.Generic;
 using System.Reflection;
@@ -7,25 +9,25 @@ using UnityEditor;
 using UnityEngine;
 using UnityEngine.UIElements;
 
-namespace SaintsField.Editor.Drawers.Wwise.GetBankDrawer
+namespace SaintsField.Editor.Drawers.Wwise.WwiseAutoGetterDrawer
 {
-    public partial class GetBankAttributeDrawer
+    public partial class WwiseAutoGetterAttributeDrawer
     {
-        private static string NameHelpBox(SerializedProperty property) => $"{property.propertyPath}__GetBank";
-
-        protected override VisualElement CreateBelowUIToolkit(SerializedProperty property, ISaintsAttribute saintsAttribute, int index,
-            VisualElement container, FieldInfo info, object parent)
-        {
-            return new HelpBox("", HelpBoxMessageType.Error)
-            {
-                style =
-                {
-                    flexGrow = 1,
-                    display = DisplayStyle.None,
-                },
-                name = NameHelpBox(property),
-            };
-        }
+        // private static string NameHelpBox(SerializedProperty property) => $"{property.propertyPath}__WwiseAutoGetter";
+        //
+        // protected override VisualElement CreateBelowUIToolkit(SerializedProperty property, ISaintsAttribute saintsAttribute, int index,
+        //     VisualElement container, FieldInfo info, object parent)
+        // {
+        //     return new HelpBox("", HelpBoxMessageType.Error)
+        //     {
+        //         style =
+        //         {
+        //             flexGrow = 1,
+        //             display = DisplayStyle.None,
+        //         },
+        //         name = NameHelpBox(property),
+        //     };
+        // }
 
         protected override void OnAwakeUIToolkit(SerializedProperty property, ISaintsAttribute saintsAttribute, int index,
             IReadOnlyList<PropertyAttribute> allAttributes, VisualElement container, Action<object> onValueChangedCallback, FieldInfo info, object parent)
@@ -35,14 +37,14 @@ namespace SaintsField.Editor.Drawers.Wwise.GetBankDrawer
             HelpBox helpBox = GetHelpBox(container, property, index);
             if (prop == null)
             {
-                helpBox.text = $"Expect Wwise.Bank, get {info.FieldType}";
+                helpBox.text = $"Expect Wwise object, get {info.FieldType}";
                 helpBox.style.display = DisplayStyle.Flex;
                 return;
             }
 
             if (prop.propertyType != SerializedPropertyType.ObjectReference)
             {
-                helpBox.text = $"Expect Wwise.Bank, get {info.FieldType}({prop.propertyType})";
+                helpBox.text = $"Expect Wwise object, get {info.FieldType}({prop.propertyType})";
                 helpBox.style.display = DisplayStyle.Flex;
                 return;
             }
@@ -51,5 +53,7 @@ namespace SaintsField.Editor.Drawers.Wwise.GetBankDrawer
         }
     }
 }
+
+#endif
 
 #endif