+++ Editor/Playa/Renderer/BaseRenderer/AbsRenderer.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using SaintsField.Editor.Drawers.ArraySizeDrawer;
-using SaintsField.Editor.Drawers.Wwise.GetBankDrawer;
+using SaintsField.Editor.Drawers.Wwise.WwiseAutoGetterDrawer;
 using SaintsField.Editor.Drawers.XPathDrawers.GetByXPathDrawer;
 using SaintsField.Editor.Playa.Utils;
 using SaintsField.Editor.Utils;
@@ -94,8 +94,9 @@ namespace SaintsField.Editor.Playa.Renderer.BaseRenderer
                     case IPlayaArraySizeAttribute arraySizeAttribute:
                         if(fieldWithInfo.SerializedProperty != null)
                         {
-                            // Debug.Log(fieldWithInfo.SerializedProperty);
-                            arraySize = fieldWithInfo.SerializedProperty.isArray
+                            // ReSharper disable once ArrangeRedundantParentheses
+                            arraySize = (fieldWithInfo.SerializedProperty.propertyType == SerializedPropertyType.Generic
+                                         && fieldWithInfo.SerializedProperty.isArray)
                                 ? GetArraySize(arraySizeAttribute, fieldWithInfo.SerializedProperty,
                                     fieldWithInfo.FieldInfo, fieldWithInfo.Target, isImGui)
                                 : (-1, -1);
@@ -179,13 +180,13 @@ namespace SaintsField.Editor.Playa.Renderer.BaseRenderer
                     return (min, max);
                 }
 #if WWISE_2024_OR_LATER || WWISE_2023_OR_LATER || WWISE_2022_OR_LATER || WWISE_2021_OR_LATER || WWISE_2020_OR_LATER || WWISE_2019_OR_LATER || WWISE_2018_OR_LATER || WWISE_2017_OR_LATER || WWISE_2016_OR_LATER || SAINTSFIELD_WWISE && !SAINTSFIELD_WWISE_DISABLE
-                case GetBankAttribute _:
+                case GetWwiseAttribute _:
                 {
                     if (!_getByXPathKeepUpdate)
                     {
                         return (-1, -1);
                     }
-                    _getByXPathKeepUpdate = GetBankAttributeDrawer.HelperGetArraySize(property, info, isImGui);
+                    _getByXPathKeepUpdate = WwiseAutoGetterAttributeDrawerHelper.HelperGetArraySize(property, info, isImGui);
                 }
                     return (-1, -1);
 #endif